map struct finish && texture struct finish

+++ export/printf/includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: sperron <sperron@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/24 09:12:40 by sperron           #+#    #+#             */
-/*   Updated: 2024/11/18 08:46:11 by sperron          ###   ########.fr       */
+/*   Updated: 2024/11/18 15:30:42 by sperron          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 
 int		ft_printf(const char *format, ...);
 int		ft_dprintf(int fd, const char *format, ...);
-int		ft_puthexa(char form, long double nbr, int fd);
+int		ft_puthexa(char form, unsigned int nbr, int fd);
 int		pf_putnbr_unsigned(unsigned int n, int fd);
 int		pf_char(int c, int fd);
 int		pf_str(char *str, int fd);
